CFS-17: Move Auth components to web part from packages
