+++ apps/web/package.json
@@ -11,12 +11,15 @@
     "check-types": "tsc --noEmit"
   },
   "dependencies": {
+    "@hookform/resolvers": "^5.2.1",
     "@repo/ui": "workspace:*",
     "lucide-react": "^0.475.0",
     "next": "^15.4.2",
     "next-themes": "^0.4.4",
     "react": "^19.1.0",
-    "react-dom": "^19.1.0"
+    "react-dom": "^19.1.0",
+    "react-hook-form": "^7.62.0",
+    "zod": "^3.25.76"
   },
   "devDependencies": {
     "@repo/eslint-config": "workspace:*",
