+++ packages/ui/package.json
@@ -8,7 +8,6 @@
     "./postcss.config": "./postcss.config.mjs",
     "./lib/*": "./src/lib/*.ts",
     "./components/*": "./src/components/*.tsx",
-    "./custom/*": "./src/custom/*.tsx",
     "./hooks/*": "./src/hooks/*.ts"
   },
   "scripts": {
